COM-1480 fix from pr comments
